Phase 1 Proxy Rewrite
Reorganized modules, and renamed files/classes so as to not confuse
python import pipeline

Staged a MCUUID class for later, this will hopefully allow us to conform
and standardize UUIDs since they are currently type fluid.

Reimplemented previously rolled back enhancements.

Added a TRACE level for logs, this can be toggled in wrapper.properties
and is defaulted to false. This logging level will trace out all packet
data feeding through the proxy. I highly recommend using this only for
low level proxy work since it produces a high volume of logging.
