+++ src/mcuuid.py
@@ -3,6 +3,10 @@
 import uuid
 
 class MCUUID (uuid.UUID):
+"""
+This class is currently not being used, but may be benefitial in regards to
+conforming UUIDs
+"""
 
     def __init__(self, hex=None, bytes=None, bytes_le=None, fields=None, int=None, version=None):
         super(MCUUID, self).__init__(hex, bytes, bytes_le, fields, int, version)
@@ -10,11 +14,3 @@ class MCUUID (uuid.UUID):
     @property
     def string(self):
         return str(self)
-
-h = MCUUID('12345678123456781234567812345678')
-
-print h.string
-print h.hex
-print h.int
-
-print uuid.UUID(bytes="12345678123456781234567812345678".decode("hex"))
