+++ src/proxy/client.py
@@ -136,7 +136,7 @@ class Client:
     def close(self):
         self.abort = True
         try:
-            self.socket.close()
+            self.socket.close() # Is this necessary?
         except Exception as e:
             self.log.error("Could not close client socket! (%s)" % e)
         if self.server:
@@ -179,7 +179,6 @@ class Client:
 
     def parse(self, pkid):  # server - bound parse ("Client" class connection)
         if pkid == 0x00 and self.state != 3:  # 0x00 is a 1.9 gameplay packet of "spawn object"
-            self.log.debug("Parsing client.............................. 0x00 and state not 3")
             if self.state == 0:   # Handshake
                 data = self.read("varint:version|string:address|ushort:port|varint:state")
                 self.version = data["version"]
@@ -204,6 +203,7 @@ class Client:
                     self.state = data["state"]
                 else:
                     self.disconnect("Invalid state '%d'" % data["state"])
+                self.log.trace("(PROXY CLIENT) -> Parsed 0x00 packet with client state 0 (HANDSHAKE):\n%s" % data)
                 return False
             elif self.state == 1:
                 sample = []
@@ -255,16 +255,17 @@ class Client:
                     self.send(0x02, "string|string", (str(self.uuid), self.username))
                     self.state = 3
                     self.log.info("%s logged in (IP: %s)" % (self.username, self.addr[0]))
+                self.log.trace("(PROXY CLIENT) -> Parsed 0x00 packet with client state 2:\n%s" % data)
                 return False
 
         if pkid == 0x01:
-            self.log.debug("Parsing client.............................. 0x01")
             # Moved 'if state == 3' out and created the if pkid == self.pktSB.CHAT_MESSAGE
             if self.state == 4:  # Encryption Response Packet
                 if self.wrapper.server.protocolVersion < 6:
                     data = self.read("bytearray_short:shared_secret|bytearray_short:verify_token")
                 else:
                     data = self.read("bytearray:shared_secret|bytearray:verify_token")
+                self.log.trace("(PROXY CLIENT) -> Parsed 0x01 packet with client state 4 (ENCRYPTION RESPONSE):\n%s" % data)
 
                 sharedSecret = encryption.decrypt_shared_secret(data["shared_secret"], self.privateKey)
                 verifyToken = encryption.decrypt_shared_secret(data["verify_token"], self.privateKey)
@@ -355,11 +356,12 @@ class Client:
             elif self.state == 5: # ping packet during status request  (What is state 5?)
                 keepAlive = self.read("long:keepAlive")["keepAlive"]
                 self.send(0x01, "long", (keepAlive,))
+                self.log.trace("(PROXY CLIENT) -> Parsed 0x01 packet with client state 5 (PING)")
                 pass
 
         if pkid == self.pktSB.CHAT_MESSAGE and self.state == 3:
-            self.log.debug("Parsing client.............................. chat_message and state 3")
             data = self.read("string:message")
+            self.log.trace("(PROXY CLIENT) -> Parsed CHAT_MESSAGE packet with client state 3:\n%s" % data)
             if data is None:
                 return False
             try:
@@ -391,29 +393,29 @@ class Client:
                 return False
             except Exception as e:
                 self.log.error("Formulating CHAT_MESSAGE failed (%s)" % e)
-                self.log.error(traceback.format_exc())
+                self.log.getTraceback()
 
             # if self.getPlayerObject().hasGroup("test"):
             #     pass
 
         if pkid == self.pktSB.PLAYER_POSITION: # player position
-            self.log.debug("Parsing client.............................. player_position")
             data = self.read("double:x|double:y|double:z|bool:on_ground")
             self.position = (data["x"], data["y"], data["z"])
+            self.log.trace("(PROXY CLIENT) -> Parsed PLAYER_POSITION packet:\n%s" % data)
         
         if pkid == self.pktSB.PLAYER_POSLOOK: # player position and look
-            self.log.debug("Parsing client.............................. player_poslook")
             data = self.read("double:x|double:y|double:z|float:yaw|float:pitch|bool:on_ground")
             self.position = (data["x"], data["y"], data["z"])
             self.head = (data["yaw"], data["pitch"])
+            self.log.trace("(PROXY CLIENT) -> Parsed PLAYER_POSLOOK packet:\n%s" % data)
             if self.server.state != 3:
                 return False
 
         if pkid == self.pktSB.PLAYER_LOOK: # Player Look
-            self.log.debug("Parsing client.............................. player_look")
             data = self.read("float:yaw|float:pitch|bool:on_ground")
             yaw, pitch = data["yaw"], data["pitch"]
             self.head = (yaw, pitch)
+            self.log.trace("(PROXY CLIENT) -> Parsed PLAYER_LOOK packet:\n%s" % data)
 
         if pkid == self.pktSB.PLAYER_DIGGING: # Player Block Dig
             if not self.isLocal:
@@ -421,9 +423,11 @@ class Client:
             if self.version < mcpacket.PROTOCOLv1_8START:
                 data = self.read("byte:status|int:x|ubyte:y|int:z|byte:face")
                 position = (data["x"], data["y"], data["z"])
+                self.log.trace("(PROXY CLIENT) -> Parsed PLAYER_DIGGING packet:\n%s" % data)
             else:
                 data = self.read("byte:status|position:position|byte:face")
                 position = data["position"]
+                self.log.trace("(PROXY CLIENT) -> Parsed PLAYER_DIGGING packet:\n%s" % data)
             if data is None:
                 return False
             # finished digging
@@ -491,7 +495,6 @@ class Client:
                 if self.version >= mcpacket.PROTOCOLv1_9REL1:
                     data = self.read("position:Location|varint:face|varint:hand|byte:CurPosX|byte:CurPosY|byte:CurPosZ")
                     hand = data["hand"]
-                    # print("interaction : %s" % str(data))
                 else:
                     data = self.read("position:Location|byte:face|slot:item|byte:CurPosX|byte:CurPosY|byte:CurPosZ")
                     helditem = data["item"]
@@ -505,6 +508,8 @@ class Client:
             # all variables populated for all versions for:
             # position, face, helditem, hand, and all three cursor positions
             # (x, y, z)
+
+            self.log.trace("(PROXY CLIENT) -> Parsed PLAYER_BLOCK_PLACEMENT packet:\n%s" % data)
 
             if face == 0:  # Compensate for block placement coordinates
                 position = (position[0], position[1] - 1, position[2])
@@ -543,7 +548,7 @@ class Client:
             if self.isLocal is not True:
                 return True
             data = self.read("rest:pack")
-            # print(data)
+            self.log.trace("(PROXY CLIENT) -> Parsed USE_ITEM packet:\n%s" % data)
             player = self.getPlayerObject()
             position = self.lastplacecoords
             helditem = player.getHeldItem()
@@ -561,6 +566,7 @@ class Client:
 
         if pkid == self.pktSB.HELD_ITEM_CHANGE: # Held Item Change
             slot = self.read("short:short")["short"]
+            self.log.trace("(PROXY CLIENT) -> Parsed HELD_ITEM_CHANGE packet:\n%s" % slot)
             if self.slot > -1 and self.slot < 9:
                 self.slot = slot
             else:
@@ -578,6 +584,7 @@ class Client:
             l3 = data["line3"]
             l4 = data["line4"]
             payload = self.wrapper.callEvent("player.createsign", {"player": self.getPlayerObject(), "position": position, "line1": l1, "line2": l2, "line3": l3, "line4": l4})
+            self.log.trace("(PROXY CLIENT) -> Parsed PLAYER_UPDATE_SIGN packet:\n%s" % data)
             if not payload:
                 return False
             if type(payload) == dict:
@@ -593,25 +600,26 @@ class Client:
             return False
         
         if pkid == self.pktSB.CLIENT_SETTINGS: # read Client Settings
-            self.log.debug("Parsing client.............................. client_settings")
             if self.version < mcpacket.PROTOCOL_1_9START:
                 data = self.read("string:locale|byte:view_distance|byte:chat_mode|bool:chat_colors|ubyte:displayed_skin_parts")
             else:
                 data = self.read("string:locale|byte:view_distance|varint:chat_mode|bool:chat_colors|ubyte:displayed_skin_parts|varint:main_hand")
             self.clientSettings = data
+            self.log.trace("(PROXY CLIENT) -> Parsed CLIENT_SETTINGS packet:\n%s" % data)
         
         if pkid == self.pktSB.CLICK_WINDOW: # click window
             data = self.read("ubyte:wid|short:slot|byte:button|short:action|byte:mode|slot:clicked")
             data['player'] = self.getPlayerObject()
+            self.log.trace("(PROXY CLIENT) -> Parsed CLICK_WINDOW packet:\n%s" % data)
             if not self.wrapper.callEvent("player.slotClick", data):
                 return False
         
         if pkid == self.pktSB.SPECTATE: # Spectate - convert packet to local server UUID
             data = self.read("uuid:target_player")
+            self.log.trace("(PROXY CLIENT) -> Parsed SPECTATE packet:\n%s" % data)
             for client in self.proxy.clients:
                 if data["target_player"].hex == client.uuid.hex:
-                    # Convert SPECTATE packet...
-                    self.server.send(self.pktSB.SPECTATE, "uuid", [client.serverUUID])
+                    self.server.send(self.pktSB.SPECTATE, "uuid", [client.serverUUID]) # Convert SPECTATE packet...
                     return False
 
         return True # Default case
@@ -626,14 +634,16 @@ class Client:
                     pkid, original = self.packet.grabPacket()
                     self.original = original
                 except EOFError as eof:
-                    self.log.error("Packet EOF (%s)" % eof)
-                    self.log.error(traceback.format_exc())
+                    # This error is often erroneous since socket data recv length is 0 when transmit ends
+                    self.log.warn("Client Packet EOF (%s)" % eof)
+                    self.log.getTraceback()
                     self.close()
                     break
                 except Exception as e:
                     if Config.debug:
+                        # Bad file descriptor often occurs, cause is currently unknown, but seemingly harmless
                         self.log.error("Failed to grab packet [CLIENT] (%s):" % e)
-                        self.log.error(traceback.format_exc())
+                        self.log.getTraceback()
                     self.close()
                     break
                 # DISABLED until github #5 is resolved
@@ -671,6 +681,6 @@ class Client:
                 if self.parse(pkid) and self.server:
                     if self.server.state == 3:
                         self.server.sendRaw(original)
-        except Exception as e:
-            self.log.error("Error in the Client->Server method (%s):" % e)
-            self.log.error(traceback.format_exc())
+        except Exception as ex:
+            self.log.error("Error in the [Client] -> [Server] handle (%s):" % ex)
+            self.log.getTraceback()
