+++ src/proxy/server.py
@@ -34,7 +34,7 @@ class Server:
         self.safe = False
         self.eid = None
 
-        # Determine packet set to use
+        # Determine packet set to use (backwards compatibility)
         if self.version >= mcpacket.PROTOCOLv1_9REL1:
             self.pktSB = mcpacket.ServerBound19
             self.pktCB = mcpacket.ClientBound19
@@ -66,10 +66,10 @@ class Server:
         if Config.debug:
             usernameofplayer = "unk"
             try:
-                usernameofplayer = str(self.client.username) # Why was this being error handled?
+                usernameofplayer = str(self.client.username) # Why is this being error handled?
             except Exception as e:
                 self.log.error("Something weird happened (%s, %s)" % (e, usernameofplayer))
-            self.log.debug("Last packet IDs (Server->Client) of player %s before disconnection: \n%s" % (usernameofplayer, str(self.lastPacketIDs)))
+            self.log.debug("Last packet IDs (Server -> Client) of player %s before disconnection: \n%s" % (usernameofplayer, str(self.lastPacketIDs)))
         self.abort = True
         self.packet = None
         try:
@@ -128,33 +128,35 @@ class Server:
             message = self.read("string:string")
             self.log.info("Disconnected from server: %s" % message["string"])
             self.client.disconnect(message)
+            self.log.trace("(PROXY SERVER) -> Parsed 0x00 packet with server state < 3")
             return False
 
         # handle keep alive packets - DISABLED until github #5 is resolved
-        # elif pkid == self.pktCB.KEEP_ALIVE and self.state == 3:
+        # if pkid == self.pktCB.KEEP_ALIVE and self.state == 3:
         #     if self.client.version > 7:
         #         pkid = self.read("varint:i")["i"]
         #         if pkid is not None:
         #             self.send(self.pktSB.KEEP_ALIVE, "varint", (pkid,))
+        #     self.log.trace("(PROXY SERVER) -> Parsed KEEP_ALIVE packet with server state 3 (ACTIVE)")
         #     return False
 
         if pkid == 0x01 and self.state == 2:
-            self.log.debug("Parsing.............................. 0x01 and state 2")
-            # This is throwing a malformed json, this should be a json string
+            # This is throwing a malformed json exception when online mode is set to true, this should be a json string
             self.client.disconnect("Server is online mode. Please turn it off in server.properties. Wrapper.py will handle authentication on its own, so do not worry about hackers.")
+            self.log.trace("(PROXY SERVER) -> Parsed 0x01 packet with server state 2 (LOGIN)")
             return False
 
         if pkid == 0x02 and self.state == 2: # Login Success - UUID & Username are sent in this packet
-            self.log.debug("Parsing.............................. 0x02 and state 2")
             self.state = 3
+            self.log.trace("(PROXY SERVER) -> Parsed 0x02 packet with server state 2 (LOGIN)")
             return False
 
         if pkid == self.pktCB.JOIN_GAME and self.state == 3:
-            self.log.debug("Parsing..............................join_game and state 3")
             if self.version < mcpacket.PROTOCOL_1_9_1_PRE:
                 data = self.read("int:eid|ubyte:gamemode|byte:dimension|ubyte:difficulty|ubyte:max_players|string:level_type")
             else:
                 data = self.read("int:eid|ubyte:gamemode|int:dimension|ubyte:difficulty|ubyte:max_players|string:level_type")
+            self.log.trace("(PROXY SERVER) -> Parsed JOIN_GAME packet with server state 3 (ACTIVE):\n%s" % data)
             oldDimension = self.client.dimension
             self.client.gamemode = data["gamemode"]
             self.client.dimension = data["dimension"]
@@ -192,6 +194,7 @@ class Server:
             position = rawdata["position"]
             try:
                 data = json.loads(rawstring)
+                self.log.trace("(PROXY SERVER) -> Parsed CHAT_MESSAGE packet with server state 3 (ACTIVE):\n%s" % data)
             except Exception as e:
                 return
 
@@ -203,22 +206,19 @@ class Server:
                     self.client.send(self.pktCB.CHAT_MESSAGE, "string|byte", (chatmsg, position))
                     return False
                 elif type(payload) == str:  # return a string-only object
-                    self.log.debug("player.Chatbox return payload sent as string")
                     self.client.send(self.pktCB.CHAT_MESSAGE, "string|byte", (payload, position))
                     return False
                 return True
             else:
                 return False
 
-            # print("The payload \n '%s' \n was never typed because it is %s" % (payload, type(payload)))
             try:
-                if data["translate"] == "chat.type.admin": # Why was this being error handled?
+                if data["translate"] == "chat.type.admin": # Why is this being error handled?
                     return False
             except Exception as e:
-                self.log.error("Something happened (%s)" % e)
+                self.log.error("Something odd happened (%s)" % e)
 
-        if pkid == 0x03 and self.state == 2:  # Set Compression:
-            self.log.debug("Parsing.............................. 0x03 and state 2")
+        if pkid == 0x03 and self.state == 2:  # Set Compression
             data = self.read("varint:threshold")
             if data["threshold"] != -1:
                 self.packet.compression = True
@@ -226,55 +226,55 @@ class Server:
             else:
                 self.packet.compression = False
                 self.packet.compressThreshold = -1
+            self.log.trace("(PROXY SERVER) -> Parsed 0x03 packet with server state 2 (LOGIN):\n%s" % data)
             return False
 
         if self.state < 3:
             return True  # remaining packets are parsed solely per "play" state
 
         if pkid == self.pktCB.TIME_UPDATE:
-            self.log.debug("Parsing.............................. time_update")
             data = self.read("long:worldage|long:timeofday")
             self.wrapper.server.timeofday = data["timeofday"]
+            self.log.trace("(PROXY SERVER) -> Parsed TIME_UPDATE packet:\n%s" % data)
             return True
 
         if pkid == self.pktCB.SPAWN_POSITION:  # Spawn Position
-            self.log.debug("Parsing.............................. spawn_position")
             data = self.read("position:spawn")
             self.wrapper.server.spawnPoint = data["spawn"]
+            self.log.trace("(PROXY SERVER) -> Parsed SPAWN_POSITION packet:\n%s" % data)
             return True
 
         if pkid == self.pktCB.RESPAWN:  # Respawn Packet
-            self.log.debug("Parsing.............................. respawn")
             data = self.read("int:dimension|ubyte:difficulty|ubyte:gamemode|level_type:string")
             self.client.gamemode = data["gamemode"]
             self.client.dimension = data["dimension"]
+            self.log.trace("(PROXY SERVER) -> Parsed RESPAWN packet:\n%s" % data)
             return True
 
         if pkid == self.pktCB.PLAYER_POSLOOK:  # Player Position and Look
-            self.log.debug("Parsing.............................. player_poslook")
             data = self.read("double:x|double:y|double:z|float:yaw|float:pitch")
             x, y, z, yaw, pitch = data["x"], data["y"], data["z"], data["yaw"], data["pitch"]
             self.client.position = (x, y, z)
+            self.log.trace("(PROXY SERVER) -> Parsed PLAYER_POSLOOK packet:\n%s" % data)
             return True
 
         if pkid == self.pktCB.USE_BED:  # Use Bed
-            self.log.debug("Parsing.............................. use_bed")
             data = self.read("varint:eid|position:location")
+            self.log.trace("(PROXY SERVER) -> Parsed USE_BED packet:\n%s" % data)
             if data["eid"] == self.eid:
                 self.client.send(self.pktCB.USE_BED, "varint|position", (self.client.eid, data["location"]))
                 return False
             return True
 
-        if pkid == self.pktCB.ANIMATION:  # self.pktCB.ANIMATION: # Animation
-            self.log.debug("Parsing.............................. animation")
+        if pkid == self.pktCB.ANIMATION: # Animation
             data = self.read("varint:eid|ubyte:animation")
+            self.log.trace("(PROXY SERVER) -> Parsed ANIMATION packet:\n%s" % data)
             if data["eid"] == self.eid:
                 self.client.send(self.pktCB.ANIMATION, "varint|ubyte", (self.client.eid, data["animation"]))
                 return False
             return True
 
         if pkid == self.pktCB.SPAWN_PLAYER:  # Spawn Player
-            self.log.debug("Parsing.............................. spawn_player")
             if self.version < mcpacket.PROTOCOL_1_9START:
                 data = self.read("varint:eid|uuid:uuid|int:x|int:y|int:z|byte:yaw|byte:pitch|short:item|rest:metadata")
                 if data["item"] < 0: # A negative Current Item crashes clients (just in case)
@@ -292,6 +292,7 @@ class Server:
                         data["item"],
                         data["metadata"])
                     )
+                self.log.trace("(PROXY SERVER) -> Parsed SPAWN_PLAYER packet:\n%s" % data)
                 return False
             else:
                 data = self.read("varint:eid|uuid:uuid|int:x|int:y|int:z|byte:yaw|byte:pitch|rest:metadata")
@@ -307,11 +308,11 @@ class Server:
                         data["pitch"],
                         data["metadata"])
                     )
+                    self.log.trace("(PROXY SERVER) -> Parsed SPAWN_PLAYER packet:\n%s" % data)
                     return False
             return True
 
         if pkid == self.pktCB.SPAWN_OBJECT:  # self.pktCB.SPAWN_OBJECT and self.state >= 3: # Spawn Object
-            self.log.debug("Parsing.............................. spawn_object")
             if self.version < mcpacket.PROTOCOL_1_9START:
                 data = self.read("varint:eid|byte:type_|int:x|int:y|int:z|byte:pitch|byte:yaw")
                 entityuuid = None
@@ -319,13 +320,13 @@ class Server:
                 data = self.read("varint:eid|uuid:objectUUID|byte:type_|int:x|int:y|int:z|byte:pitch|byte:yaw|int:info|short:velocityX|short:velocityY|short:velocityZ")
                 entityuuid = data["objectUUID"]
             eid, type_, x, y, z, pitch, yaw = data["eid"], data["type_"], data["x"], data["y"], data["z"], data["pitch"], data["yaw"]
+            self.log.trace("(PROXY SERVER) -> Parsed SPAWN_OBJECT packet:\n%s" % data)
             if not self.wrapper.server.world:
                 return
             self.wrapper.server.world.entities[data["eid"]] = Entity(eid, entityuuid, type_, (x, y, z), (pitch, yaw), True)
             return True
 
         if pkid == self.pktCB.SPAWN_MOB:  # Spawn Mob
-            self.log.debug("Parsing.............................. spawn_mob")
             if self.version < mcpacket.PROTOCOL_1_9START:
                 data = self.read("varint:eid|ubyte:type_|int:x|int:y|int:z|byte:pitch|byte:yaw|byte:head_pitch|short:velocityX|short:velocityY|short:velocityZ|rest:metadata")
                 entityuuid = None
@@ -333,6 +334,7 @@ class Server:
                 data = self.read("varint:eid|uuid:entityUUID|ubyte:type_|int:x|int:y|int:z|byte:pitch|byte:yaw|byte:head_pitch|short:velocityX|short:velocityY|short:velocityZ|rest:metadata")
                 entityuuid = data["entityUUID"]
             eid, type_, x, y, z, pitch, yaw, head_pitch = data["eid"], data["type_"], data["x"], data["y"], data["z"], data["pitch"], data["yaw"], data["head_pitch"]
+            self.log.trace("(PROXY SERVER) -> Parsed SPAWN_MOB packet:\n%s" % data)
             if not self.wrapper.server.world:
                 return
             # this will need entity UUID's added at some point
@@ -343,18 +345,18 @@ class Server:
                 # Temporary! These packets need to be filtered for cross-server stuff.
                 return True
             data = self.read("varint:eid|byte:dx|byte:dy|byte:dz")
+            self.log.trace("(PROXY SERVER) -> Parsed ENTITY_RELATIVE_MOVE packet:\n%s" % data)
             if not self.wrapper.server.world:
                 return
             if self.wrapper.server.world.getEntityByEID(data["eid"]) is not None:
                 self.wrapper.server.world.getEntityByEID(data["eid"]).moveRelative((data["dx"], data["dy"], data["dz"]))
-            # self.log.debug("Parsing.............................. entity_relative_move %s, %s" % (data, self.wrapper.server.world.getEntityByEID(data["eid"])))
 
         if pkid == self.pktCB.ENTITY_TELEPORT:  # Entity Teleport
-            # self.log.debug("Parsing.............................. entity_teleport")
             if self.version < mcpacket.PROTOCOLv1_8START:
                 # Temporary! These packets need to be filtered for cross-server stuff.
                 return True
             data = self.read("varint:eid|int:x|int:y|int:z|byte:yaw|byte:pitch")
+            self.log.trace("(PROXY SERVER) -> Parsed ENTITY_TELEPORT packet:\n%s" % data)
             if not self.wrapper.server.world:
                 return
             if self.wrapper.server.world.getEntityByEID(data["eid"]) is not None:
@@ -362,23 +364,23 @@ class Server:
         
         if pkid == self.pktCB.ENTITY_HEAD_LOOK:
             data = self.read("varint:eid|byte:angle")
-            # self.log.debug("Parsing.............................. entity_head_look (%s)" % data)
+            self.log.trace("(PROXY SERVER) -> Parsed ENTITY_HEAD_LOOK packet:\n%s" % data)
         
         if pkid == self.pktCB.ENTITY_STATUS:  # Entity Status
-            # self.log.debug("Parsing.............................. entity_status")
             if self.version < mcpacket.PROTOCOLv1_8START:
                 # Temporary! These packets need to be filtered for cross-server stuff.
                 return True
             data = self.read("int:eid|byte:status")
+            self.log.trace("(PROXY SERVER) -> Parsed ENTITY_STATUS packet:\n%s" % data)
         
         if pkid == self.pktCB.ATTACH_ENTITY:  # Attach Entity
-            # self.log.debug("Parsing.............................. attach_entity")
             if self.version < mcpacket.PROTOCOLv1_8START:
                 # Temporary! These packets need to be filtered for cross-server stuff.
                 return True
             data = self.read("varint:eid|varint:vid|bool:leash")
             eid, vid, leash = data["eid"], data["vid"], data["leash"]
             player = self.getPlayerByEID(eid)
+            self.log.trace("(PROXY SERVER) -> Parsed ATTACH_ENTITY packet:\n%s" % data)
             if player is None:
                 return
             if eid == self.eid:
@@ -396,60 +398,63 @@ class Server:
                     return False
         
         if pkid == self.pktCB.ENTITY_METADATA:  # Entity Metadata
-            # self.log.debug("Parsing.............................. entity_metadata")
             if self.version < mcpacket.PROTOCOLv1_8START:
                 # Temporary! These packets need to be filtered for cross-server stuff.
                 return True
             data = self.read("varint:eid|rest:metadata")
+            self.log.trace("(PROXY SERVER) -> Parsed ENTITY_METADATA packet:\n%s" % data)
             if data["eid"] == self.eid:
                 self.client.send(self.pktCB.ENTITY_METADATA,"varint|raw", (self.client.eid, data["metadata"]))
                 return False
         
-        if pkid == self.pktCB.ENTITY_EFFECT:  # Entity Effect
-            # self.log.debug("Parsing.............................. entity_effect")
+        if pkid == self.pktCB.ENTITY_EFFECT: # Entity Effect
             if self.version < mcpacket.PROTOCOLv1_8START:
                 # Temporary! These packets need to be filtered for cross-server stuff.
                 return True
             data = self.read("varint:eid|byte:effect_id|byte:amplifier|varint:duration|bool:hide")
+            self.log.trace("(PROXY SERVER) -> Parsed ENTITY_EFFECT packet:\n%s" % data)
             if data["eid"] == self.eid:
                 self.client.send(self.pktCB.ENTITY_EFFECT, "varint|byte|byte|varint|bool", (self.client.eid, data["effect_id"], data["amplifier"], data["duration"], data["hide"]))
                 return False
         
-        if pkid == self.pktCB.REMOVE_ENTITY_EFFECT:  # Remove Entity Effect
-            # self.log.debug("Parsing.............................. remove_entity_effect")
+        if pkid == self.pktCB.REMOVE_ENTITY_EFFECT: # Remove Entity Effect
             if self.version < mcpacket.PROTOCOLv1_8START:
                 # Temporary! These packets need to be filtered for cross-server stuff.
                 return True
             data = self.read("varint:eid|byte:effect_id")
+            self.log.trace("(PROXY SERVER) -> Parsed REMOVE_ENTITY_EFFECT packet:\n%s" % data)
             if data["eid"] == self.eid:
                 self.client.send(self.pktCB.REMOVE_ENTITY_EFFECT, "varint|byte", (self.client.eid, data["effect_id"]))
                 return False
         
         if pkid == self.pktCB.ENTITY_PROPERTIES:  # Entity Properties
-            # self.log.debug("Parsing.............................. entity_properties")
             if self.version < mcpacket.PROTOCOLv1_8START:
                 # Temporary! These packets need to be filtered for cross-server stuff.
                 return True
             data = self.read("varint:eid|rest:properties")
+            self.log.trace("(PROXY SERVER) -> Parsed ENTITY_PROPERTIES packet:\n%s" % data)
             if data["eid"] == self.eid:
                 self.client.send(self.pktCB.ENTITY_PROPERTIES, "varint|raw", (self.client.eid, data["properties"]))
                 return False
 
         # if pkid == self.pktCB.CHUNK_DATA: # Chunk Data
-        # if self.client.packet.compressThreshold == -1:
-        #   self.log.debug("Client compression enabled, setting to 256")
-        #   self.client.packet.setCompression(256)
+        #     if self.client.packet.compressThreshold == -1:
+        #         self.log.debug("Client compression enabled, setting to 256")
+        #         self.client.packet.setCompression(256)
+        #     self.log.trace("(PROXY SERVER) -> Parsed CHUNK_DATA packet")
+
 
         # if self.pktCB.BLOCK_CHANGE: # Block Change - disabled - not doing anything at this point
         #     if self.version < mcpacket.PROTOCOLv1_8START:
         #         # Temporary! These packets need to be filtered for cross-server stuff.
         #         return True
         #     data = self.read("position:location|varint:pkid")
+        #     self.log.trace("(PROXY SERVER) -> Parsed BLOCK_CHANGE packet:\n%s" % data)
         
         if pkid == self.pktCB.MAP_CHUNK_BULK: # Map Chunk Bulk (no longer exists in 1.9)
-            self.log.debug("Parsing.............................. map_chunk_bulk")
             if self.version > mcpacket.PROTOCOLv1_8START and self.version < mcpacket.PROTOCOL_1_9START:
                 data = self.read("bool:skylight|varint:chunks")
+                self.log.trace("(PROXY SERVER) -> Parsed MAP_CHUNK_BULK packet:\n%s" % data)
                 for i in xrange(data["chunks"]):
                     meta = self.read("int:x|int:z|ushort:primary")
                     bitmask = bin(meta["primary"])[2:].zfill(16)
@@ -465,34 +470,31 @@ class Server:
                         else:
                             # Null Chunk
                             chunkColumn += bytearray(16 * 16 * 16 * 2)
-                #self.wrapper.server.world.setChunk(meta["x"], meta["z"], world.Chunk(chunkColumn, meta["x"], meta["z"]))
-                # print "Reading chunk %d,%d" % (meta["x"], meta["z"])
         
         if pkid == self.pktCB.CHANGE_GAME_STATE:  # Change Game State
-            self.log.debug("Parsing.............................. change_game_state")
             data = self.read("ubyte:reason|float:value")
             if data["reason"] == 3:
                 self.client.gamemode = data["value"]
+            self.log.trace("(PROXY SERVER) -> Parsed CHANGE_GAME_STATE packet:\n%s" % data)
         
         if pkid == self.pktCB.SET_SLOT:  # Set Slot
-            self.log.debug("Parsing.............................. set_slot")
             if self.version < mcpacket.PROTOCOLv1_8START:
                 # Temporary! These packets need to be filtered for cross-server stuff.
                 return True
             data = self.read("byte:wid|short:slot|slot:data")
             if data["wid"] == 0:
                 self.client.inventory[data["slot"]] = data["data"]
+            self.log.trace("(PROXY SERVER) -> Parsed SET_SLOT packet:\n%s" % data)
 
         # if pkid == 0x30: # Window Items
         #   data = self.read("byte:wid|short:count")
-        #   print data["count"]
         #   if data["wid"] == 0:
         #       for slot in range(1, data["count"]):
         #           data = self.read("slot:data")
         #           self.client.inventory[slot] = data["data"]
+        #   self.log.trace("(PROXY SERVER) -> Parsed 0x30 packet:\n%s" % data)
 
         if pkid == self.pktCB.PLAYER_LIST_ITEM:  # player list item
-            self.log.debug("Parsing.............................. player_list_item")
             if self.version > mcpacket.PROTOCOLv1_8START:
                 head = self.read("varint:action|varint:length")
                 z = 0
@@ -514,11 +516,11 @@ class Server:
                         properties = playerclient.properties
                         raw = ""
                         for prop in properties:
-                            raw += self.client.packet.send_string(prop["name"]) # name
-                            raw += self.client.packet.send_string(prop["value"]) # value
+                            raw += self.client.packet.send_string(prop["name"])
+                            raw += self.client.packet.send_string(prop["value"])
                             if "signature" in prop:
                                 raw += self.client.packet.send_bool(True)
-                                raw += self.client.packet.send_string(prop["signature"]) # signature
+                                raw += self.client.packet.send_string(prop["signature"])
                             else:
                                 raw += self.client.packet.send_bool(False)
                         raw += self.client.packet.send_varInt(0)
@@ -527,14 +529,17 @@ class Server:
                         self.client.send(self.pktCB.PLAYER_LIST_ITEM, "varint|varint|uuid|string|varint|raw", (0, 1, playerclient.uuid, playerclient.username, len(properties), raw))
                     elif head["action"] == 1:
                         data = self.read("varint:gamemode")
+                        self.log.trace("(PROXY SERVER) -> Parsed PLAYER_LIST_ITEM packet:\n%s" % data)
                         self.client.send(self.pktCB.PLAYER_LIST_ITEM, "varint|varint|uuid|varint", (1, 1, uuid, data["gamemode"]))
                     elif head["action"] == 2:
                         data = self.read("varint:ping")
+                        self.log.trace("(PROXY SERVER) -> Parsed PLAYER_LIST_ITEM packet:\n%s" % data)
                         self.client.send(self.pktCB.PLAYER_LIST_ITEM, "varint|varint|uuid|varint", (2, 1, uuid, data["ping"]))
                     elif head["action"] == 3:
                         data = self.read("bool:has_display")
                         if data["has_display"]:
                             data = self.read("string:displayname")
+                            self.log.trace("(PROXY SERVER) -> Parsed PLAYER_LIST_ITEM packet:\n%s" % data)
                             self.client.send(self.pktCB.PLAYER_LIST_ITEM, "varint|varint|uuid|bool|string", (3, 1, uuid, True, data["displayname"]))
                         else:
                             self.client.send(self.pktCB.PLAYER_LIST_ITEM, "varint|varint|uuid|varint", (3, 1, uuid, False))
@@ -543,13 +548,13 @@ class Server:
                     return False
 
         if pkid == self.pktCB.DISCONNECT:  # Disconnect
-            self.log.debug("Parsing.............................. disconnect")
             message = self.read("json:json")["json"]
             self.log.info("Disconnected from server: %s" % message)
             if not self.client.isLocal:
                 self.close()
             else:
                 self.client.disconnect(message)
+            self.log.trace("(PROXY SERVER) -> Parsed DISCONNECT packet")
             return False
 
         return True # Default case
@@ -565,14 +570,16 @@ class Server:
                             del self.lastPacketIDs[i]
                             break
                 except EOFError as eof:
+                    # This error is often erroneous since socket data recv length is 0 when transmit ends
                     self.log.error("Packet EOF (%s)" % eof)
-                    self.log.error(traceback.format_exc())
+                    self.log.getTraceback()
                     self.close()
                     break
                 except Exception as e1:
                     if Config.debug:
+                        # Bad file descriptor often occurs, cause is currently unknown, but seemingly harmless
                         self.log.error("Failed to grab packet [SERVER] (%s):" % e1)
-                        self.log.error(traceback.format_exc())
+                        self.log.getTraceback()
                     return
                 if self.client.abort:
                     self.close()
@@ -581,502 +588,6 @@ class Server:
                     self.client.sendRaw(original)
         except Exception as e2:
             if Config.debug:
-                self.log.error("Error in the Server->Client method (%s):" % e2)
-                self.log.error(traceback.format_exc())
+                self.log.error("Error in the [SERVER] -> [CLIENT] handle (%s):" % e2)
+                self.log.getTraceback()
             self.close()
-
-# from mcpkt import ServerBound18 as defPacketsSB
-# from mcpkt import ClientBound18 as defPacketsCB
-# from mcpkt import ServerBound19 as PacketsSB19
-# from mcpkt import ClientBound19 as PacketsCB19
-
-# PROTOCOL_1_9_1_PRE = 108 # post- 1.9 "pre releases (1.9.1 pre-3 and later
-# PROTOCOLv1_9REL1 = 107  # start of stable 1.9 release (or most current snapshop that is documented by protocol)
-# PROTOCOL_1_9START = 48  # start of 1.9 snapshots
-# PROTOCOLv1_8START = 6
-
-# class Server: # Handle Server Connection  ("client bound" packets)
-#     def __init__(self, client, wrapper, ip=None, port=None):
-#         self.client = client
-#         self.wrapper = wrapper
-#         self.ip = ip
-#         self.port = port
-#         self.abort = False
-#         self.isServer = True
-#         self.proxy = wrapper.proxy
-#         self.lastPacketIDs = []
-
-#         self.state = 0 # 0 = init, 1 = motd, 2 = login, 3 = active, 4 = authorizing
-#         self.packet = None
-#         self.version = self.wrapper.server.protocolVersion
-#         self.log = wrapper.log
-#         self.safe = False
-#         self.eid = None
-        
-#         # Determine packet set to use
-#         self.pktSB = defPacketsSB
-#         self.pktCB = defPacketsCB
-#         if self.version >= PROTOCOLv1_9REL1:
-#             self.pktSB = PacketsSB19
-#             self.pktCB = PacketsCB19
-            
-#     def connect(self):
-#         self.socket = socket.socket()
-#         if self.ip == None:
-#             self.socket.connect(("localhost", self.wrapper.config["Proxy"]["server-port"]))
-#         else:
-#             self.socket.connect((self.ip, self.port))
-#             self.client.isLocal = False
-
-#         self.packet = Packet(self.socket, self)
-#         self.packet.version = self.client.version
-#         self.username = self.client.username
-
-#         self.send = self.packet.send
-#         self.read = self.packet.read
-#         self.sendRaw = self.packet.sendRaw
-
-#         t = threading.Thread(target=self.flush, args=())
-#         t.daemon = True
-#         t.start()
-#     def close(self, reason="Disconnected", kill_client=True):
-#         if Config.debug:
-#             usernameofplayer = "unk"
-#             try:
-#                 usernameofplayer = str(self.client.username)
-#             except:
-#                 pass
-#             print("Last packet IDs (Server->Client) of player %s before disconnection: \n%s\n" %
-#                   (usernameofplayer, str(self.lastPacketIDs)))
-#             # print self.lastPacketIDs
-#         self.abort = True
-#         self.packet = None
-#         try:
-#             self.socket.close()
-#         except:
-#             pass
-#         if self.client.isLocal == False and kill_client:  # Ben's cross-server hack
-#             self.client.isLocal = True
-#             self.client.send(self.pktCB.changegamestate, "ubyte|float", (1, 0))  # "end raining"
-#             self.client.send(self.pktCB.chatmessage, "string|byte", ("{text:'Disconnected from server: %s', color:red}" % reason.replace("'", "\\'"), 0))
-#             self.client.connect()
-#             return
-
-#         # I may remove this later so the client can remain connected upon server disconnection
-# #       self.client.send(0x02, "string|byte", (json.dumps({"text": "Disconnected from server. Reason: %s" % reason, "color": "red"}),0))
-# #       self.abort = True
-# #       self.client.connect()
-#         if kill_client:
-#             self.client.abort = True
-#             self.client.server = None
-#             self.client.close()
-#     def getPlayerByEID(self, eid):
-#         for client in self.wrapper.proxy.clients:
-#             try:
-#                 if client.server.eid == eid: return self.getPlayerContext(client.username)
-#             except: print("client.server.eid failed!\nserverEid: %s\nEid: %s" %
-#                           (str(client.server.eid), str(eid)))
-#         return False
-#     def getPlayerContext(self, username):
-#         try: return self.wrapper.server.players[username]
-#         except: return False
-#     def flush(self):
-#         while not self.abort:
-#             self.packet.flush()
-#         #   try:
-# #               self.packet.flush()
-# #           except:
-# #               print "Error while flushing, stopping"
-# #               print traceback.format_exc()
-# #               self.close()
-# #               break
-#             time.sleep(0.03)
-#     def parse(self, id, original):  # client - bound parse ("Server" class connection)
-#         """
-#         Client-bound "Server Class"
-#         """
-#         # disconnect, I suppose...
-#         if id == 0x00 and self.state < 3:
-#             message = self.read("string:string")
-#             self.log.info("Disconnected from server: %s" % message["string"])
-#             self.client.disconnect(message)
-#             return False
-            
-#         # handle keep alive packets 
-#         if False == True: #id == self.pktCB.keepalive and self.state == 3:
-#                 if self.client.version > 7:
-#                     id = self.read("varint:i")["i"]
-#                     if not id == None:
-#                         self.send(self.pktSB.keepalive, "varint", (id,))
-#                 return False
-                
-#         if id == 0x01 and self.state == 2:
-#             self.client.disconnect("Server is online mode. Please turn it off in server.properties.\n\nWrapper.py will handle authentication on its own, so do not worry about hackers.")
-#             return False
-
-#         # Login Success - UUID & Username are sent in this packet
-#         if id == 0x02 and self.state == 2:
-#             self.state = 3
-#             return False
-
-#         if id == self.pktCB.joingame and self.state == 3:
-#             if self.version < PROTOCOL_1_9_1_PRE:
-#                 data = self.read("int:eid|ubyte:gamemode|byte:dimension|ubyte:difficulty|ubyte:max_players|string:level_type")
-#             elif self.version >= PROTOCOL_1_9_1_PRE:
-#                 data = self.read("int:eid|ubyte:gamemode|int:dimension|ubyte:difficulty|ubyte:max_players|string:level_type")
-#             oldDimension = self.client.dimension
-#             self.client.gamemode = data["gamemode"]
-#             self.client.dimension = data["dimension"]
-#             self.eid = data["eid"]  # This is the EID of the player on this particular server - not always the EID that the client is aware of
-#             if self.client.handshake:
-#                 dimensions = [-1, 0, 1]
-#                 if oldDimension == self.client.dimension:
-#                     for l in dimensions:
-#                         if l != oldDimension:
-#                             dim = l
-#                             break
-#                     self.client.send(self.pktCB.respawn, "int|ubyte|ubyte|string", (l, data["difficulty"], data["gamemode"], data["level_type"]))
-#                 self.client.send(self.pktCB.respawn, "int|ubyte|ubyte|string", (self.client.dimension, data["difficulty"], data["gamemode"], data["level_type"]))
-#                 #self.client.send(0x01, "int|ubyte|byte|ubyte|ubyte|string|bool", (self.eid, self.client.gamemode, self.client.dimension, data["difficulty"], data["max_players"], data["level_type"], False))
-#                 self.eid = data["eid"]
-#                 self.safe = True
-#                 return False
-#             else:
-#                 self.client.eid = data["eid"]
-#                 self.safe = True
-#             self.client.handshake = True
-
-#             #print "Sending change game state packet..."
-#             self.client.send(self.pktCB.changegamestate, "ubyte|float", (3, self.client.gamemode))
-#             if UNIVERSAL_CONNECT is True:
-#                 clientversion = self.packet.version
-#                 serverversion = self.wrapper.server.protocolVersion
-#                 if clientversion < PROTOCOL_1_9_1_PRE <= serverversion:
-#                     self.client.send(self.pktCB.joingame, "int|ubyte|byte|ubyte|ubyte|string",
-#                                      (data["eid"], data["gamemode"], data["dimension"], data["difficulty"], data["max_players"], data["level_type"]))
-#                     return False
-#         if id == self.pktCB.chatmessage and self.state == 3:
-#             rawdata = self.read("string:json|byte:position")
-#             rawstring = rawdata["json"]
-#             position = rawdata["position"]
-#             try:
-#                 data = json.loads(rawstring)
-#             except: return
-
-#             ## added code
-#             payload = self.wrapper.callEvent("player.chatbox", {"player": self.client.getPlayerObject(), "json": data})
-#             if payload is False:
-#                 return False
-
-#             if type(payload) == dict:  # return a "chat" protocol formatted dictionary http://wiki.vg/Chat
-#                 chatmsg = json.dumps(payload)
-#                 self.client.send(self.pktCB.chatmessage, "string|byte", (chatmsg, position))
-#                 return False
-
-#             if type(payload) == str:  # return a string-only object
-#                 print("player.Chatbox return payload sent as string")
-#                 self.client.send(self.pktCB.chatmessage, "string|byte", (payload, position))
-#                 return False
-
-#             if payload is True:
-#                 return True
-
-#             #print("The payload \n '%s' \n was never typed because it is %s" % (payload, type(payload)))
-#             try:
-#                 if data["translate"] == "chat.type.admin": return False
-#             except: pass
-
-#         if id == 0x03 and self.state == 2: # Set Compression:
-#             data = self.read("varint:threshold")
-#             if data["threshold"] != -1:
-#                 self.packet.compression = True
-#                 self.packet.compressThreshold = data["threshold"]
-#             else:
-#                 self.packet.compression = False
-#                 self.packet.compressThreshold = -1
-#             return False
-
-#         if self.state < 3:return True #  remaining packets are parsed solely per "play" state
-
-#         if id == self.pktCB.timeupdate:
-#             data = self.read("long:worldage|long:timeofday")
-#             self.wrapper.server.timeofday = data["timeofday"]
-#             return True
-#         if id == self.pktCB.spawnposition: # Spawn Position
-#             data = self.read("position:spawn")
-#             self.wrapper.server.spawnPoint = data["spawn"]
-#             return True
-#         if id == self.pktCB.respawn: # Respawn Packet
-#             data = self.read("int:dimension|ubyte:difficulty|ubyte:gamemode|level_type:string")
-#             self.client.gamemode = data["gamemode"]
-#             self.client.dimension = data["dimension"]
-#             return True
-#         if id == self.pktCB.playerposlook: # Player Position and Look
-#             data = self.read("double:x|double:y|double:z|float:yaw|float:pitch")
-#             x, y, z, yaw, pitch = data["x"], data["y"], data["z"], data["yaw"], data["pitch"]
-#             self.client.position = (x, y, z)
-#             return True
-#         if id == self.pktCB.usebed: # Use Bed
-#             data = self.read("varint:eid|position:location")
-#             if data["eid"] == self.eid:
-#                 self.client.send(self.pktCB.usebed, "varint|position", (self.client.eid, data["location"]))
-#                 return False
-#             return True
-#         if id == self.pktCB.animation: #  #self.pktCB.Animation: # Animation
-#             data = self.read("varint:eid|ubyte:animation")
-#             if data["eid"] == self.eid:
-#                 self.client.send(self.pktCB.animation, "varint|ubyte", (self.client.eid, data["animation"]))
-#                 return False
-#             return True
-#         if id == self.pktCB.spawnplayer:  # Spawn Player
-#             if self.version < PROTOCOL_1_9START:
-#                 data = self.read("varint:eid|uuid:uuid|int:x|int:y|int:z|byte:yaw|byte:pitch|short:item|rest:metadata")
-#                 if data["item"] < 0: data["item"] = 0  # A negative Current Item crashes clients (just in ncase)
-#                 clientserverid = self.proxy.getClientByServerUUID(data["uuid"])
-#                 if clientserverid:
-#                     self.client.send(self.pktCB.spawnplayer, "varint|uuid|int|int|int|byte|byte|short|raw", (
-#                         data["eid"],
-#                         clientserverid.uuid,
-#                         data["x"],
-#                         data["y"],
-#                         data["z"],
-#                         data["yaw"],
-#                         data["pitch"],
-#                         data["item"],
-#                         data["metadata"]))
-#                 return False
-#             else:
-#                 data = self.read("varint:eid|uuid:uuid|int:x|int:y|int:z|byte:yaw|byte:pitch|rest:metadata")
-#                 clientserverid = self.proxy.getClientByServerUUID(data["uuid"])
-#                 if clientserverid:
-#                     self.client.send(self.pktCB.spawnplayer, "varint|uuid|int|int|int|byte|byte|raw", (
-#                         data["eid"],
-#                         clientserverid.uuid,
-#                         data["x"],
-#                         data["y"],
-#                         data["z"],
-#                         data["yaw"],
-#                         data["pitch"],
-#                         data["metadata"]))
-#                     return False
-#             return True
-#         if id == self.pktCB.spawnobject: #  #self.pktCB.spawnobject and self.state >= 3: # Spawn Object
-#             if self.version < PROTOCOL_1_9START:
-#                 data = self.read("varint:eid|byte:type_|int:x|int:y|int:z|byte:pitch|byte:yaw")
-#                 entityuuid = None
-#             else:
-#                 data = self.read("varint:eid|uuid:objectUUID|byte:type_|int:x|int:y|int:z|byte:pitch|byte:yaw|int:info|"
-#                                  "short:velocityX|short:velocityY|short:velocityZ")
-#                 entityuuid = data["objectUUID"]
-#             eid, type_, x, y, z, pitch, yaw = data["eid"], data["type_"], data["x"], data["y"], data["z"], data["pitch"], data["yaw"]
-#             if not self.wrapper.server.world: return
-#             self.wrapper.server.world.entities[data["eid"]] = Entity(eid, entityuuid, type_, (x, y, z), (pitch, yaw), True)
-#             return True
-#         if id == self.pktCB.spawnmob:  # Spawn Mob
-#             if self.version < PROTOCOL_1_9START:
-#                 data = self.read("varint:eid|ubyte:type_|int:x|int:y|int:z|byte:pitch|byte:yaw|byte:head_pitch|short:velocityX|short:velocityY|short:velocityZ|rest:metadata")
-#                 entityuuid = None
-#             else:
-#                 data = self.read("varint:eid|uuid:entityUUID|ubyte:type_|int:x|int:y|int:z|byte:pitch|byte:yaw|byte:head_pitch|short:velocityX|short:velocityY|short:velocityZ|rest:metadata")
-#                 entityuuid = data["entityUUID"]
-#             eid, type_, x, y, z, pitch, yaw, head_pitch = data["eid"], data["type_"], data["x"], data["y"], data["z"], data["pitch"], data["yaw"], data["head_pitch"]
-#             if not self.wrapper.server.world: return
-#             # this will need entity UUID's added at some point
-#             self.wrapper.server.world.entities[data["eid"]] = Entity(eid, entityuuid, type_, (x, y, z), (pitch, yaw, head_pitch), False)
-#         if id == self.pktCB.entityrelativemove:  # Entity Relative Move
-#             if self.version < PROTOCOLv1_8START: return True # Temporary! These packets need to be filtered for cross-server stuff.
-#             data = self.read("varint:eid|byte:dx|byte:dy|byte:dz")
-#             if not self.wrapper.server.world: return
-#             if not self.wrapper.server.world.getEntityByEID(data["eid"]) == None:
-#                 self.wrapper.server.world.getEntityByEID(data["eid"]).moveRelative((data["dx"], data["dy"], data["dz"]))
-#         if id == self.pktCB.entityteleport:  # Entity Teleport
-#             if self.version < PROTOCOLv1_8START: return True # Temporary! These packets need to be filtered for cross-server stuff.
-#             data = self.read("varint:eid|int:x|int:y|int:z|byte:yaw|byte:pitch")
-#             if not self.wrapper.server.world: return
-#             if not self.wrapper.server.world.getEntityByEID(data["eid"]) == None:
-#                 self.wrapper.server.world.getEntityByEID(data["eid"]).teleport((data["x"], data["y"], data["z"]))
-#         if id == self.pktCB.entityheadlook:
-#             data = self.read("varint:eid|byte:angle")
-#         if id == self.pktCB.entitystatus:  # Entity Status
-#             if self.version < PROTOCOLv1_8START: return True # Temporary! These packets need to be filtered for cross-server stuff.
-#             data = self.read("int:eid|byte:status")
-#         if id == self.pktCB.attachentity:  # Attach Entity
-#             if self.version < PROTOCOLv1_8START: return True # Temporary! These packets need to be filtered for cross-server stuff.
-#             data = self.read("varint:eid|varint:vid|bool:leash")
-#             eid, vid, leash = data["eid"], data["vid"], data["leash"]
-#             player = self.getPlayerByEID(eid)
-#             if player == None: return
-#             if eid == self.eid:
-#                 if vid == -1:
-#                     self.wrapper.callEvent("player.unmount", {"player": player})
-#                     self.client.riding = None
-#                 else:
-#                     self.wrapper.callEvent("player.mount", {"player": player, "vehicle_id": vid, "leash": leash})
-#                     if not self.wrapper.server.world: return
-#                     self.client.riding = self.wrapper.server.world.getEntityByEID(vid)
-#                     self.wrapper.server.world.getEntityByEID(vid).rodeBy = self.client
-#                 if eid != self.client.eid:
-#                     self.client.send(self.pktCB.attachentity, "varint|varint|bool", (self.client.eid, vid, leash))
-#                     return False
-#         if id == self.pktCB.entitymetadata:  # Entity Metadata
-#             if self.version < PROTOCOLv1_8START: return True # Temporary! These packets need to be filtered for cross-server stuff.
-#             data = self.read("varint:eid|rest:metadata")
-#             if data["eid"] == self.eid:
-#                 self.client.send(self.pktCB.entitymetadata, "varint|raw", (self.client.eid, data["metadata"]))
-#                 return False
-#         if id == self.pktCB.entityeffect:  # Entity Effect
-#             if self.version < PROTOCOLv1_8START: return True # Temporary! These packets need to be filtered for cross-server stuff.
-#             data = self.read("varint:eid|byte:effect_id|byte:amplifier|varint:duration|bool:hide")
-#             if data["eid"] == self.eid:
-#                 self.client.send(self.pktCB.entityeffect, "varint|byte|byte|varint|bool", (self.client.eid, data["effect_id"], data["amplifier"], data["duration"], data["hide"]))
-#                 return False
-#         if id == self.pktCB.removeentityeffect:  # Remove Entity Effect
-#             if self.version < PROTOCOLv1_8START: return True # Temporary! These packets need to be filtered for cross-server stuff.
-#             data = self.read("varint:eid|byte:effect_id")
-#             if data["eid"] == self.eid:
-#                 self.client.send(self.pktCB.removeentityeffect, "varint|byte", (self.client.eid, data["effect_id"]))
-#                 return False
-#         if id == self.pktCB.entityproperties:  # Entity Properties
-#             if self.version < PROTOCOLv1_8START: return True # Temporary! These packets need to be filtered for cross-server stuff.
-#             data = self.read("varint:eid|rest:properties")
-#             if data["eid"] == self.eid:
-#                 self.client.send(self.pktCB.entityproperties, "varint|raw", (self.client.eid, data["properties"]))
-#                 return False
-
-#     #   if id == self.pktCB.chunkdata: # Chunk Data
-# #           if self.client.packet.compressThreshold == -1:
-# #               print "CLIENT COMPRESSION ENABLED"
-# #               self.client.packet.setCompression(256)
-
-#         if True == False: #  #self.pktCB.blockchange: # Block Change - disabled - not doing anything at this point
-#             if self.version < PROTOCOLv1_8START: return True # Temporary! These packets need to be filtered for cross-server stuff.
-#             data = self.read("position:location|varint:id")
-
-#         if id == self.pktCB.mapchunkbulk: # Map Chunk Bulk (no longer exists in 1.9)
-#             if self.version > PROTOCOLv1_8START and self.version < PROTOCOL_1_9START:
-#                 data = self.read("bool:skylight|varint:chunks")
-#                 chunks = []
-#                 for i in range(data["chunks"]):
-#                     meta = self.read("int:x|int:z|ushort:primary")
-#                     chunks.append(meta)
-#                 for i in range(data["chunks"]):
-#                     meta = chunks[i]
-#                     bitmask = bin(meta["primary"])[2:].zfill(16)
-#                     primary = []
-#                     for i in bitmask:
-#                         if i == "0": primary.append(False)
-#                         if i == "1": primary.append(True)
-#                     chunkColumn = bytearray()
-#                     for i in primary:
-#                         if i == True:
-#                             chunkColumn += bytearray(self.packet.read_data(16*16*16 * 2)) # packetanisc
-#                             if self.client.dimension == 0:
-#                                 metalight = bytearray(self.packet.read_data(16*16*16))
-#                             if data["skylight"]:
-#                                 skylight = bytearray(self.packet.read_data(16*16*16))
-#                         else:
-#                             chunkColumn += bytearray(16*16*16 * 2) # Null Chunk
-#                 #self.wrapper.server.world.setChunk(meta["x"], meta["z"], world.Chunk(chunkColumn, meta["x"], meta["z"]))
-#                 #print "Reading chunk %d,%d" % (meta["x"], meta["z"])
-#         if id == self.pktCB.changegamestate: # Change Game State
-#             data = self.read("ubyte:reason|float:value")
-#             if data["reason"] == 3:
-#                 self.client.gamemode = data["value"]
-#         if id == self.pktCB.setslot: # Set Slot
-#             if self.version < PROTOCOLv1_8START: return True # Temporary! These packets need to be filtered for cross-server stuff.
-#             data = self.read("byte:wid|short:slot|slot:data")
-#             if data["wid"] == 0:
-#                 self.client.inventory[data["slot"]] = data["data"]
-#         # if id == 0x30: # Window Items
-#         #   data = self.read("byte:wid|short:count")
-#         #   print data["count"]
-#         #   if data["wid"] == 0:
-#         #       for slot in range(1, data["count"]):
-#         #           data = self.read("slot:data")
-#         #           self.client.inventory[slot] = data["data"]
-#         if id == self.pktCB.playerlistitem:  # player list item
-#             if self.version > PROTOCOLv1_8START:
-#                 head = self.read("varint:action|varint:length")
-#                 z = 0
-#                 while z < head["length"]:
-#                     serverUUID = self.read("uuid:uuid")["uuid"]
-#                     playerclient = self.client.proxy.getClientByServerUUID(serverUUID)
-#                     if not playerclient:
-#                         z += 1
-#                         continue
-#                     try: uuid = playerclient.uuid
-#                     except:
-#                         # uuid = playerclient
-#                         print("playercleint.uuid failed in playerlist item")
-#                         z += 1
-#                         continue
-#                     z += 1
-#                     if head["action"] == 0:
-#                         properties = playerclient.properties
-#                         raw = ""
-#                         for i in properties:
-#                             raw += self.client.packet.send_string(i["name"]) # name
-#                             raw += self.client.packet.send_string(i["value"]) # value
-#                             if "signature" in i:
-#                                 raw += self.client.packet.send_bool(True)
-#                                 raw += self.client.packet.send_string(i["signature"]) # signature
-#                             else:
-#                                 raw += self.client.packet.send_bool(False)
-#                         raw += self.client.packet.send_varInt(0)
-#                         raw += self.client.packet.send_varInt(0)
-#                         raw += self.client.packet.send_bool(False)
-#                         self.client.send(self.pktCB.playerlistitem, "varint|varint|uuid|string|varint|raw", (0, 1, playerclient.uuid, playerclient.username, len(properties), raw))
-#                     elif head["action"] == 1:
-#                         data = self.read("varint:gamemode")
-#                         self.client.send(self.pktCB.playerlistitem, "varint|varint|uuid|varint", (1, 1, uuid, data["gamemode"]))
-#                     elif head["action"] == 2:
-#                         data = self.read("varint:ping")
-#                         self.client.send(self.pktCB.playerlistitem, "varint|varint|uuid|varint", (2, 1, uuid, data["ping"]))
-#                     elif head["action"] == 3:
-#                         data = self.read("bool:has_display")
-#                         if data["has_display"]:
-#                             data = self.read("string:displayname")
-#                             self.client.send(self.pktCB.playerlistitem, "varint|varint|uuid|bool|string", (3, 1, uuid, True, data["displayname"]))
-#                         else:
-#                             self.client.send(self.pktCB.playerlistitem, "varint|varint|uuid|varint", (3, 1, uuid, False))
-#                     elif head["action"] == 4:
-#                         self.client.send(self.pktCB.playerlistitem, "varint|varint|uuid", (4, 1, uuid))
-#                     return False
-#         if id == self.pktCB.disconnect: # disconnect
-#             message = self.read("json:json")["json"]
-#             self.log.info("Disconnected from server: %s" % message)
-#             if self.client.isLocal == False:
-#                 self.close()
-#             else:
-#                 self.client.disconnect(message)
-#             return False
-#         return True
-#     def handle(self):
-#         try:
-#             while not self.abort:
-#                 try:
-#                     id, original = self.packet.grabPacket()
-#                     self.lastPacketIDs.append((hex(id), len(original)))
-#                     if len(self.lastPacketIDs) > 10:
-#                         for i,v in enumerate(self.lastPacketIDs):
-#                             del self.lastPacketIDs[i]
-#                             break
-#                 except EOFError:
-#                     print traceback.format_exc()
-#                     self.close()
-#                     break
-#                 except:
-#                     if Config.debug:
-#                         print "Failed to grab packet (SERVER)"
-#                         print traceback.format_exc()
-#                     return
-#                 if self.client.abort:
-#                     self.close()
-#                     break
-#                 if self.parse(id, original) and self.safe:
-#                     self.client.sendRaw(original)
-#         except:
-#             if Config.debug:
-#                 print "Error in the Server->Client method:"
-#                 print traceback.format_exc()
-#             self.close()
